% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.athena_operations.R
\name{batch_get_query_execution}
\alias{batch_get_query_execution}
\title{Returns the details of a single query execution or a list of up to 50 query executions, which you provide as an array of query execution ID strings}
\usage{
batch_get_query_execution(QueryExecutionIds)
}
\arguments{
\item{QueryExecutionIds}{An array of query execution IDs.}
}
\description{
Returns the details of a single query execution or a list of up to 50 query executions, which you provide as an array of query execution ID strings. To get a list of query execution IDs, use ListQueryExecutions. Query executions are different from named (saved) queries. Use BatchGetNamedQuery to get details about named queries.
}
\examples{

}
