% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.athena_operations.R
\name{create_named_query}
\alias{create_named_query}
\title{Creates a named query}
\usage{
create_named_query(Name, Description = NULL, Database, QueryString,
  ClientRequestToken = NULL)
}
\arguments{
\item{Name}{[required] The plain language name for the query.}

\item{Description}{A brief explanation of the query.}

\item{Database}{[required] The database to which the query belongs.}

\item{QueryString}{[required] The text of the query itself. In other words, all query statements.}

\item{ClientRequestToken}{A unique case-sensitive string used to ensure the request to create the query is idempotent (executes only once). If another \code{CreateNamedQuery} request is received, the same response is returned and another query is not created. If a parameter has changed, for example, the \code{QueryString}, an error is returned.

This token is listed as not required because AWS SDKs (for example the AWS SDK for Java) auto-generate the token for users. If you are not using the AWS SDK or the AWS CLI, you must provide this token or the action will fail.}
}
\description{
Creates a named query.
}
\details{
For code samples using the AWS SDK for Java, see \href{http://docs.aws.amazon.com/athena/latest/ug/code-samples.html}{Examples and Code Samples} in the \emph{Amazon Athena User Guide}.
}
\section{Accepted Parameters}{
\preformatted{create_named_query(
  Name = "string",
  Description = "string",
  Database = "string",
  QueryString = "string",
  ClientRequestToken = "string"
)
}
}

