% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.athena_operations.R
\name{delete_named_query}
\alias{delete_named_query}
\title{Deletes a named query}
\usage{
delete_named_query(
  NamedQueryId = "string"
)
}
\arguments{
\item{NamedQueryId}{[required] The unique ID of the query to delete.}
}
\description{
Deletes a named query.
}
\details{
For code samples using the AWS SDK for Java, see \href{http://docs.aws.amazon.com/athena/latest/ug/code-samples.html}{Examples and Code Samples} in the \emph{Amazon Athena User Guide}.
}
\examples{

}
