% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.athena_operations.R
\name{get_named_query}
\alias{get_named_query}
\title{Returns information about a single query}
\usage{
get_named_query(NamedQueryId)
}
\arguments{
\item{NamedQueryId}{The unique ID of the query. Use ListNamedQueries to get query IDs.}
}
\description{
Returns information about a single query.
}
\examples{

}
