% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.athena_operations.R
\name{get_query_results}
\alias{get_query_results}
\title{Returns the results of a single query execution specified by QueryExecutionId}
\usage{
get_query_results(QueryExecutionId, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{QueryExecutionId}{[required] The unique ID of the query execution.}

\item{NextToken}{The token that specifies where to start pagination if a previous request was truncated.}

\item{MaxResults}{The maximum number of results (rows) to return in this request.}
}
\description{
Returns the results of a single query execution specified by \code{QueryExecutionId}. This request does not execute the query but returns results. Use StartQueryExecution to run a query.
}
\section{Accepted Parameters}{
\preformatted{get_query_results(
  QueryExecutionId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

