% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.athena_package.R
\docType{package}
\name{paws.athena-package}
\alias{paws.athena}
\alias{paws.athena-package}
\title{paws.athena: Amazon Athena}
\description{
Amazon Athena is an interactive query service that
lets you use standard SQL to analyze data directly in Amazon S3. You
can point Athena at your data in Amazon S3 and run ad-hoc queries and
get results in seconds. Athena is serverless, so there is no
infrastructure to set up or manage. You pay only for the queries you
run. Athena scales automatically---executing queries in parallel---so
results are fast, even with large datasets and complex queries. For
more information, see [What is Amazon
Athena](http://docs.aws.amazon.com/athena/latest/ug/what-is.html) in
the *Amazon Athena User Guide*.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
