% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.athena_operations.R
\name{stop_query_execution}
\alias{stop_query_execution}
\title{Stops a query execution}
\usage{
stop_query_execution(QueryExecutionId)
}
\arguments{
\item{QueryExecutionId}{[required] The unique ID of the query execution to stop.}
}
\description{
Stops a query execution.
}
\details{
For code samples using the AWS SDK for Java, see \href{http://docs.aws.amazon.com/athena/latest/ug/code-samples.html}{Examples and Code Samples} in the \emph{Amazon Athena User Guide}.
}
\section{Accepted Parameters}{
\preformatted{stop_query_execution(
  QueryExecutionId = "string"
)
}
}

