% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.athena_operations.R
\name{get_query_execution}
\alias{get_query_execution}
\title{Returns information about a single execution of a query}
\usage{
get_query_execution(QueryExecutionId)
}
\arguments{
\item{QueryExecutionId}{[required] The unique ID of the query execution.}
}
\description{
Returns information about a single execution of a query. Each time a query executes, information about the query execution is saved with a unique ID.
}
\section{Accepted Parameters}{
\preformatted{get_query_execution(
  QueryExecutionId = "string"
)
}
}

